[#211] blobstor: Rename Delete to DeleteBig
